refactor: 提取 useCommentData
