+++ src/components/respository/CommentSubmit.tsx
@@ -1,6 +1,7 @@
 import Image from 'next/image';
-import { FormEventHandler, useEffect, useState } from 'react';
+import { FormEventHandler } from 'react';
 
+import useCommentData from '@/hooks/useCommentData';
 import useLogin from '@/hooks/useLogin';
 import useUserInfo from '@/hooks/useUserInfo';
 
@@ -143,33 +144,6 @@ function CommentSubmit(props: {
       </div>
     </div>
   );
-}
-
-function useCommentData() {
-  const storageKey = 'COMMENT_DATA';
-  const [commentData, setCommentData] = useState({
-    comment: '',
-    isUsed: false,
-    score: 0,
-  });
-
-  useEffect(() => {
-    setCommentData(JSON.parse(localStorage.getItem(storageKey) || '{}'));
-  }, []);
-
-  const storeCommentData = (data: {
-    comment: string;
-    isUsed: boolean;
-    score: number;
-  }) => {
-    localStorage.setItem(storageKey, JSON.stringify(data));
-    setCommentData(data);
-  };
-
-  return {
-    commentData,
-    setCommentData: storeCommentData,
-  };
 }
 
 export default CommentSubmit;
